backend/movies_admin/requirements/dev.txt
-r base.txt

asgiref==3.2.10
django-debug-toolbar==2.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
gunicorn==20.0.4

